central and outer separated
